src/requirements.txt
absl-py==1.3.0
anyio==3.3.2
appnope==0.1.2
argon2-cffi==21.1.0
ariadne==0.14.1
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==21.12b0
bleach==4.1.0
brotlipy==0.7.0
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==4.2.4
certifi==2022.9.24
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1625835350093/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454858/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click==8.0.1
cloudpickle==2.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==22.9.0
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1618231726924/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494586/work
cycler==0.11.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.4.0
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==22.10.26
fonttools==4.28.5
gast==0.4.0
glob2==0.7
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gradescope-utils==0.4.0
graphql-core==3.1.7
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1624380716006/work
gym==0.23.1
gym-notices==0.0.6
h11==0.13.0
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
httplib2==0.20.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
image==1.5.33
importlib-metadata==4.11.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.1.18
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.6
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.7
notebook==6.4.4
numpy==1.23.4
oauthlib==3.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1624996792650/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pygame==2.1.2
Pygments==2.10.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291460269/work
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
regex==2022.4.24
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1624999868738/work
Send2Trash==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.1.0
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==1.10.2
torchaudio==0.11.0
torchlight==0.0.1
torchvision==0.11.3
tornado==6.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1628985621016/work
traitlets==5.1.0
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
uvicorn==0.17.1
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.14.1
yapf==0.32.0
zipp==3.8.0
